+++ ambry-clustermap/src/main/java/com.github.ambry.clustermap/CompositeClusterManager.java
@@ -114,7 +114,7 @@ class CompositeClusterManager implements ClusterMap {
   }
 
   @Override
-  public short getLocalDatacenterId() {
+  public byte getLocalDatacenterId() {
     return LEGACY_DATACENTER_ID;
   }
