+++ ambry-clustermap/src/main/java/com.github.ambry.clustermap/HelixClusterManager.java
@@ -167,7 +167,7 @@ class HelixClusterManager implements ClusterMap {
   }
 
   @Override
-  public short getLocalDatacenterId() {
+  public byte getLocalDatacenterId() {
     return LEGACY_DATACENTER_ID;
   }
