+++ ambry-clustermap/src/main/java/com.github.ambry.clustermap/StaticClusterManager.java
@@ -95,7 +95,7 @@ class StaticClusterManager implements ClusterMap {
   }
 
   @Override
-  public short getLocalDatacenterId() {
+  public byte getLocalDatacenterId() {
     return LEGACY_DATACENTER_ID;
   }
