+++ ambry-commons/src/test/java/com.github.ambry.commons/ResponseHandlerTest.java
@@ -74,7 +74,7 @@ public class ResponseHandlerTest {
     }
 
     @Override
-    public short getLocalDatacenterId() {
+    public byte getLocalDatacenterId() {
       return LEGACY_DATACENTER_ID;
     }
